[FEATURE][CEPSCLOUD][VCAN_PAGE] GUI signalizes incomplete user changes
The only feedback displayed for incomplete downstream settings was
a disabled upload button. In the case of a removal of one downstream
definition which lead to a valid setting overall didn't influence the
upload button's status at all.

Incomplete entries are colored differently form complete ones,
the upload button behaves correctly in the case of the removal of
entries.