[FEATURE][CORE][WS_API] Command EXPORTED_EVENTS added
Events are introduced via the symbol-kind syntax and therefore
not accessible by a QUERY command. Hence a new command to access
the list of defined events was in dire need. A further complication
stems from the fact that each non trivial simulation defines a
bunch of events of which the majority serves a purely technical
purpose. Only the meaningful events should be accessible by the
WS API.

We introduce the EXPORTED_EVENTS command. This command grants
access to the list of all exported events. Exported events were introduced
a long time ago (Dingo!).